Rename module names + move extract steps to ingestions folder
