+++ apps/andi/lib/andi/input_schemas/extract_http_steps.ex
@@ -1,6 +1,6 @@
 defmodule Andi.InputSchemas.ExtractHttpSteps do
   @moduledoc false
-  alias Andi.InputSchemas.Datasets.ExtractHttpStep
+  alias Andi.InputSchemas.Ingestions.ExtractHttpStep
   alias Andi.Repo
   alias Andi.InputSchemas.StructTools
