+++ apps/andi/lib/andi/input_schemas/ingestions/extract_steps/extract_header.ex
@@ -1,4 +1,4 @@
-defmodule Andi.InputSchemas.Datasets.ExtractHeader do
+defmodule Andi.InputSchemas.Ingestions.ExtractHeader do
   @moduledoc false
   use Ecto.Schema
   import Ecto.Changeset
