+++ apps/andi/lib/andi/input_schemas/ingestions/extract_steps/extract_secret_step.ex
@@ -1,4 +1,4 @@
-defmodule Andi.InputSchemas.Datasets.ExtractSecretStep do
+defmodule Andi.InputSchemas.Ingestions.ExtractSecretStep do
   @moduledoc false
   use Ecto.Schema
   import Ecto.Changeset
