+++ apps/andi/test/unit/andi/input_schemas/datasets/extract_s3_step_test.exs
@@ -1,8 +1,8 @@
-defmodule Andi.InputSchemas.Datasets.ExtractS3StepTest do
+defmodule Andi.InputSchemas.Ingestions.ExtractS3StepTest do
   use ExUnit.Case
 
-  alias Andi.InputSchemas.Datasets.ExtractS3Step
-  alias Andi.InputSchemas.Datasets.ExtractStep
+  alias Andi.InputSchemas.Ingestions.ExtractS3Step
+  alias Andi.InputSchemas.Ingestions.ExtractStep
 
   test "given changes with valid headers map, properly casts" do
     changes = %{
